LoginController, some changes, handled exception
