+++ src/main/java/com/cb/controllers/AdminController.java
@@ -75,31 +75,6 @@ public class AdminController {
         }
 
     }
-    @RequestMapping(value = "/loginuser", method = RequestMethod.POST)
-    public String loginUser(HttpServletRequest req, Map<String, String> model, @ModelAttribute("userBean") UserDAL userDAL, Model p, Model c) {
-        List<UserDAL> allEmails = userService.getUserByEmail(userDAL.getEmail());
-        List<UserDAL> allPasswords = userService.getUserByPassword(userDAL.getPassword());
-        String userName = "";
-        for (UserDAL items: allEmails
-        ) {
-            userName = items.getUserName();
-        }
-
-        if (allEmails.size() > 0 && allPasswords.size() > 0) {
-            HttpSession userSession = req.getSession();
-            userSession.setAttribute("userName", userName);
-            List<PartyDAL> partiesList = partyService.getParties();
-            List<CharacterDAL> charactersList = characterService.getCharacters();
-            p.addAttribute("partiesList", partiesList);
-            c.addAttribute("charactersList", charactersList);
-            return "createCharacter";
-
-        } else {
-            model.put("error", "User already does not exist");
-            return "index";
-
-        }
-    }
 
     /* Displays object data into form for the given id.
      * The @PathVariable puts URL data into variable.*/
